PR 2: Long vs Int and agg column names (#734, #733, #732)
- Schema: map Int32 -> IntegerType, Int64 -> LongType (was both to Long).
- Count agg: cast to Int64 for PySpark LongType (count(*) and count(col)).
- Agg alias already produces 'count' etc.; schema now reflects correct types.

+++ crates/robin-sparkless-polars/src/functions/rest.rs
@@ -6,9 +6,12 @@ use super::types::parse_type_name;
 use crate::column::Column;
 use polars::prelude::*;
 
-/// Count aggregation
+/// Count aggregation (PySpark LongType; cast to Int64 for schema parity #734).
 pub fn count(col: &Column) -> Column {
-    Column::from_expr(col.expr().clone().count(), Some("count".to_string()))
+    Column::from_expr(
+        col.expr().clone().count().cast(DataType::Int64),
+        Some("count".to_string()),
+    )
 }
 
 /// Sum aggregation
